Move plutarch-ledger-api tests to testlib
