+++ plutarch-ledger-api/plutarch-ledger-api.cabal
@@ -42,17 +42,6 @@ common common-lang
     TypeOperators
     UndecidableInstances
 
-common test-lang
-  import:        common-lang
-  ghc-options:   -O2 -threaded -rtsopts -with-rtsopts=-N
-  build-depends:
-    , plutarch
-    , plutarch-ledger-api
-    , plutus-ledger-api
-    , QuickCheck
-    , tasty
-    , tasty-quickcheck
-
 library
   import:          common-lang
   exposed-modules:
@@ -88,23 +77,3 @@ library
     , serialise
 
   hs-source-dirs:  src
-
-test-suite tests
-  import:         test-lang
-  type:           exitcode-stdio-1.0
-  main-is:        Main.hs
-  other-modules:
-    Laws
-    Regressions
-    Utils
-    V1
-    V2
-    V3
-
-  build-depends:
-    , plutarch-orphanage
-    , plutus-tx
-    , prettyprinter
-    , tasty-hunit
-
-  hs-source-dirs: test
