+++ plutarch-testlib/Plutarch/Test/Utils.hs
@@ -1,0 +1,11 @@
+{-# LANGUAGE AllowAmbiguousTypes #-}
+{-# LANGUAGE PolyKinds #-}
+
+module Plutarch.Test.Utils (
+  fewerTests,
+) where
+
+import Test.Tasty.QuickCheck (QuickCheckTests)
+
+fewerTests :: QuickCheckTests -> QuickCheckTests -> QuickCheckTests
+fewerTests divisor = (`quot` divisor)
