+++ plutarch-testlib/plutarch-testlib.cabal
@@ -1,0 +1,113 @@
+cabal-version: 3.0
+name:          plutarch-testlib
+version:       1.0.0
+license:       MIT
+
+common lang
+  default-language:   GHC2021
+  default-extensions:
+    BlockArguments
+    DataKinds
+    DefaultSignatures
+    DeriveAnyClass
+    DerivingStrategies
+    DerivingVia
+    DisambiguateRecordFields
+    DuplicateRecordFields
+    FunctionalDependencies
+    GADTs
+    ImpredicativeTypes
+    LambdaCase
+    LexicalNegation
+    LiberalTypeSynonyms
+    MonadComprehensions
+    MultiWayIf
+    NegativeLiterals
+    NoFlexibleInstances
+    NoMonomorphismRestriction
+    NoStarIsType
+    NondecreasingIndentation
+    NumDecimals
+    OverloadedLabels
+    OverloadedLists
+    OverloadedRecordDot
+    OverloadedStrings
+    PackageImports
+    PartialTypeSignatures
+    PatternSynonyms
+    QualifiedDo
+    QuantifiedConstraints
+    QuasiQuotes
+    RankNTypes
+    RecordWildCards
+    RecursiveDo
+    RoleAnnotations
+    TypeFamilies
+    TypeFamilyDependencies
+    UnicodeSyntax
+    UnliftedFFITypes
+    UnliftedNewtypes
+    ViewPatterns
+
+  build-depends:      base >=4.9 && <5
+  ghc-options:
+    -Weverything -Wno-unused-do-bind -Wno-missing-kind-signatures
+    -Wno-partial-type-signatures -Werror -Wno-implicit-prelude
+    -Wno-name-shadowing -Wno-unsafe -Wno-missing-safe-haskell-mode
+    -Wno-missing-local-signatures -Wno-prepositive-qualified-module
+    -Wno-missing-import-lists -Wno-monomorphism-restriction
+    -Wno-missed-specialisations -Wno-all-missed-specializations
+    -Wno-unrecognised-warning-flags -Wno-unrecognised-pragmas
+    -Wno-operator-whitespace -Wno-unticked-promoted-constructors
+    -fprint-equality-relations -fprint-explicit-kinds
+    -fprint-explicit-foralls -Wno-unused-packages
+
+library
+  import:          lang
+  build-depends:
+    , bytestring
+    , constraints
+    , containers
+    , cryptonite
+    , data-default
+    , flat
+    , generics-sop
+    , lens
+    , mtl
+    , plutarch
+    , plutarch-ledger-api
+    , plutarch-orphanage
+    , plutus-ledger-api
+    , plutus-tx
+    , prettyprinter
+    , QuickCheck
+    , tasty
+    , tasty-quickcheck
+
+  exposed-modules:
+    Plutarch.Test.Laws
+    Plutarch.Test.Utils
+
+test-suite tests
+  import:         lang
+  type:           exitcode-stdio-1.0
+  main-is:        Test.hs
+  other-modules:
+    Plutarch.Test.Suite.PlutarchLedgerApi.Regressions
+    Plutarch.Test.Suite.PlutarchLedgerApi.V1
+    Plutarch.Test.Suite.PlutarchLedgerApi.V2
+    Plutarch.Test.Suite.PlutarchLedgerApi.V3
+
+  build-depends:
+    , plutarch
+    , plutarch-ledger-api
+    , plutarch-orphanage
+    , plutarch-testlib
+    , plutus-ledger-api
+    , plutus-tx
+    , prettyprinter
+    , tasty
+    , tasty-hunit
+    , tasty-quickcheck
+
+  hs-source-dirs: test
