+++ plutarch-testlib/test/Plutarch/Test/Suite/PlutarchLedgerApi/V3.hs
@@ -1,9 +1,10 @@
-module V3 (tests) where
+module Plutarch.Test.Suite.PlutarchLedgerApi.V3 (tests) where
 
 import Plutarch.LedgerApi.V3 qualified as PLA
+import Plutarch.Test.Laws (checkLedgerProperties)
+import Plutarch.Test.Utils (fewerTests)
 import PlutusLedgerApi.V3.Orphans ()
 import Test.Tasty (TestTree, adjustOption, testGroup)
-import Utils (checkLedgerProperties, fewerTests)
 
 tests :: TestTree
 tests =
